Add last aditions for mongo.
