+++ go.mod
@@ -1,0 +1,7 @@
+module github.com/InterloperStudios/ims-api
+
+go 1.15
+
+require github.com/gorilla/mux v1.8.0
+
+require go.mongodb.org/mongo-driver v1.4.4
